pyproject.toml
[tool.poetry]
name = "cehckbox-vqa"
version = "1.0"
description = "checkbox detection and recognition"
authors = ["logesh0169 <logesh.b@zucisystems.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "cehckbox_vqa"}]

[tool.poetry.dependencies]
python = "^3.9"
colorlog = "^6.7.0"
pillow = "^9.4.0"
transformers = "4.25.1"
sentencepiece = "^0.1.97"
protobuf = "3.20"
torchvision = "0.10.0"
matplotlib = "^3.6.3"
keras = "2.2.5"
pandas = "^1.5.2"
albumentations = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
